Found the root cause of the redisson shutdown exception problem
The classCache and proxyCache is been set to static (class level) which
means previous test have registered the class is still been cached even
after the tests were finished. New tests fetched the old proxied class
from the cache with reference to a redisson instance that has been
created in the previous tests and that has already been shutdown when
the tests were finished.