Do not send blank passwords as AUTH commands
When using a YAML config similar to this:
```yml
singleServerConfig:
    address: "redis://${REDIS_HOST:-localhost}:${REDIS_PORT:-6379}"
    password: "${REDIS_PASSWORD}"
```
and the client tries to connect to an instance with no password set (i.e. developer install of redis listening only on localhost) the current implementation tries sending a blank password.

This change updates the password handling in Redisson so that if the password is blank, then it works the same as no password (`null`).

Signed-off-by: John J. Aylward <john.aylward@sagenet.com>